Extract base64 encoding to different file
